Make bit offset (bo) an insigned int (size_t) through out

+++ t/t_buf.h
@@ -100,5 +100,5 @@ int t_pck_read ( lua_State *luaVM, struct t_pck *p, const unsigned char *buffer)
 int t_pck_write( lua_State *luaVM, struct t_pck *p, unsigned char *buffer );
 
 // helpers for the Packers
-struct t_pck *t_pck_getpck( lua_State *luaVM, int pos, int *bo );
+struct t_pck *t_pck_getpck( lua_State *luaVM, int pos, size_t *bo );
 int           lt_pcr__call( lua_State *luaVM );
